Moved the Moonpay to a SFSafariVC
